requirements.txt
PyYAML~=6.0
transmission-rpc~=3.3.0
qbittorrent-api~=2022.1.27